+++ libkc3/buf.c
@@ -1197,7 +1197,7 @@ sw buf_read_until_str_into_buf (s_buf *buf, const s_str *end,
 sw buf_read_until_str_into_file (s_buf *buf, const s_str *end,
                                    s_str *path)
 {
-  s32 fd;
+  s64 fd;
   s_buf dest = {0};
   sw r;
   if (! file_open_w(path, &fd)) {
